+++ nbproject/private/private.xml
@@ -3,11 +3,11 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
+            <file>file:/C:/Apache2.4/htdocs/AccountManagement/public_html/PHP/Reader_Editor.php</file>
             <file>file:/C:/Apache2.4/htdocs/AccountManagement/public_html/AccountCreation.html</file>
             <file>file:/C:/Apache2.4/htdocs/AccountManagement/public_html/useraccount.html</file>
             <file>file:/C:/Apache2.4/htdocs/AccountManagement/public_html/PHP/Validation.php</file>
             <file>file:/C:/Apache2.4/htdocs/AccountManagement/public_html/PHP/dbConnect.php</file>
-            <file>file:/C:/Apache2.4/htdocs/AccountManagement/public_html/css/stylesheet.css</file>
             <file>file:/C:/Apache2.4/htdocs/AccountManagement/public_html/index.html</file>
             <file>file:/C:/Apache2.4/htdocs/AccountManagement/public_html/login.html</file>
         </group>
