+++ public_html/AccountCreation.html
@@ -73,7 +73,7 @@
           <div class="container">
             <div class="row">
               <div class="col-md-12">
-                <form data-toggle="validator" role="form" action="PHP/Validation.php" method="POST">
+                <form data-toggle="validator" role="form" action="PHP/Reader_Editor.php" method="POST">
                   <div class="form-group">
                     <label form="inputName" class="control-label" >
                       First Name
