+++ public_html/PHP/dbConnect.php
@@ -13,5 +13,5 @@
         die("Connection failed: ".$conn->connect_error);
     }
     
-    echo "Connected Successfully  - $dbhost";
+   // echo "Connected Successfully  - $dbhost";
 ?>
