+++ public_html/login.html
@@ -41,7 +41,7 @@
       <div class="row-fluid">
         <div class="span12 appMain">
           
-          <form action="PHP/Validation.php" method="POST" role="form" style="display: block;">
+          <form action="PHP/Reader_Editor.php" method="POST" role="form" style="display: block;">
             <div class="form-group">
               <input type="text" name="username" id="username"  class="form-control" placeholder="Username" value="">
             </div>
