+++ public_html/useraccount.html
@@ -53,12 +53,7 @@
                         <div class="container">
                             <div class="row">
                                 <div class="col-md-12">
-                                    <!-- content of page -->
-                                    http://stackoverflow.com/questions/9874291/get-the-values-of-form-array-with-jquery
-                                    <?php 
-  
-    echo json_encode($FormResults);    
-?>
+                                  <h1>This Can contain stats about the webpage</h1>
                                 </div>
                             </div>
                         </div>
